requirements.txt
setuptools~=69.5.1
fastapi==0.109.2
uvicorn==0.27.0
pydantic==2.7.1
pymongo~=4.7.2
python-dotenv~=1.0.1
gql~=3.5.0
requests~=2.32.2

space-exploration-common-lib==0.0.13